fixed cascade deleting
