+++ models/prompts.py
@@ -1,31 +1,7 @@
-from sqlalchemy import Integer, Column, String, DateTime, func, JSON, Table, ForeignKey, UniqueConstraint, MetaData
+from sqlalchemy import Integer, Column, String, DateTime, func, JSON
 from sqlalchemy.orm import relationship, backref
 
 from tools import db_tools, db, rpc_tools
-
-
-# prompts_tags = Table(
-#     'models_prompts_tags_association', MetaData(schema="tenant"),
-#     Column('tag_id', Integer, ForeignKey('tenant.models_prompts_tags.id', ondelete="CASCADE")),
-#     Column('prompt_id', Integer, ForeignKey('tenant.models_prompts.id', ondelete="CASCADE")),
-# )
-
-class PromptTag(db_tools.AbstractBaseMixin, db.Base):
-    __tablename__ = "models_prompts_tags_association"
-    __table_args__ = {'schema': 'tenant'}
-
-    id = Column(Integer, primary_key=True)
-    prompt_id = Column(Integer, ForeignKey('tenant.models_prompts.id', ondelete='CASCADE'))
-    tag_id = Column(Integer, ForeignKey('tenant.models_prompts_tags.id', ondelete='CASCADE'))
-
-
-class Tag(db_tools.AbstractBaseMixin, db.Base):
-    __tablename__ = "models_prompts_tags"
-    __table_args__ = {'schema': 'tenant'}
-
-    id = Column(Integer, primary_key=True)
-    tag = Column(String(250), unique=True, nullable=False)
-    color = Column(String(15))
 
 
 class Prompt(
@@ -45,4 +21,4 @@ class Prompt(
     created_at = Column(DateTime, nullable=False, server_default=func.now())
     updated_at = Column(DateTime, onupdate=func.now())
     tags = relationship("Tag", secondary='tenant.models_prompts_tags_association', 
-        backref=backref("prompts"), cascade="all,delete")
+        backref=backref("prompts"))
